refactor: 랜덤 값에 따라 자동차 위치를 이동하는 책임을 나눔
