+++ src/main/java/racingcar/model/Cars.java
@@ -4,9 +4,6 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import camp.nextstep.edu.missionutils.Randoms;
-import racingcar.utils.Numbers;
-
 public class Cars {
 
     private final List<Car> cars;
@@ -17,28 +14,6 @@ public class Cars {
 
     public List<Car> getCars() {
         return Collections.unmodifiableList(cars);
-    }
-
-    private int randomGenerator() {
-        int randomNumber = Randoms.pickNumberInRange(
-                Numbers.RANDOM_START_NUMBER.getNumber(),
-                Numbers.RANDOM_END_NUMBER.getNumber()
-        );
-        return randomNumber;
-    }
-
-    private void moveForward() {
-        cars.stream()
-                .filter(car -> randomGenerator() >= Numbers.MOVE_POSITION_NUMBER.getNumber())
-                .forEach(car -> car.updatePosition());
-    }
-
-    public void printPlayers() {
-        moveForward();
-        cars.stream()
-            .map(car -> car.getName() + " : " + car.getPosition())
-            .forEach(System.out::println);
-        System.out.println();
     }
 
     private int getMaxLength(){
